✨ UserCard.jsx 파일 생성
